% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSymmetric.R
\name{.allPermutations}
\alias{.allPermutations}
\title{Generate all permutations of 1:n}
\usage{
.allPermutations(n)
}
\arguments{
\item{n}{Integer}
}
\value{
List of integer vectors
}
\description{
Generate all permutations of 1:n
}
\keyword{internal}
