% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symNMF.R
\name{.betaDiv}
\alias{.betaDiv}
\title{Compute Beta-divergence D_beta(X || Y)}
\usage{
.betaDiv(X, Y, beta)
}
\arguments{
\item{X}{Target matrix}

\item{Y}{Approximation matrix}

\item{beta}{Beta parameter}
}
\value{
Scalar divergence value
}
\description{
Compute Beta-divergence D_beta(X || Y)
}
\keyword{internal}
