% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HigherOrderPower.R
\name{.contractMode}
\alias{.contractMode}
\title{Contract an array along a specific mode with a vector}
\usage{
.contractMode(arr, v, mode)
}
\arguments{
\item{arr}{Multi-dimensional array}

\item{v}{Vector}

\item{mode}{Mode to contract}
}
\value{
Array with one fewer dimension
}
\description{
Contract an array along a specific mode with a vector
}
\keyword{internal}
