% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symNMF.R
\name{.muExponent}
\alias{.muExponent}
\title{MU exponent for Beta-divergence}
\usage{
.muExponent(beta)
}
\arguments{
\item{beta}{Beta parameter}
}
\value{
Exponent gamma
}
\description{
Returns the exponent gamma for the multiplicative update rule.
gamma = 1 for Beta in [1,2], otherwise adjusted for convergence guarantee.
}
\references{
Fevotte, C. & Idier, J. (2011). Neural Computation, 23(9).
}
\keyword{internal}
