% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HigherOrderPower.R
\name{.outerRepeat}
\alias{.outerRepeat}
\title{Compute repeated outer product of a vector}
\usage{
.outerRepeat(v, d)
}
\arguments{
\item{v}{Numeric vector}

\item{d}{Number of times to take outer product}
}
\value{
d-dimensional array
}
\description{
Compute repeated outer product of a vector
}
\keyword{internal}
