% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HigherOrderPower.R
\name{.tensorVectorContract}
\alias{.tensorVectorContract}
\title{Contract a tensor with a vector along all modes except the first}
\usage{
.tensorVectorContract(arr, v, d)
}
\arguments{
\item{arr}{Array of dimension (n, n, ..., n)}

\item{v}{Numeric vector of length n}

\item{d}{Number of modes}
}
\value{
Numeric vector of length n
}
\description{
Contract a tensor with a vector along all modes except the first
}
\keyword{internal}
